Generate implementation of endpoint methods
All public slots are preparing a base URI with query parameters when
needed as well as headers. After preliminary preparation, a connection
between virtual on*Request and lambda processing response is established.
Within body of lambda, a corresponding virtual on*Success or on*Error
will be invoked before final response to a user via callback method.
Implementation of POST and PUT pre-requests is not provided since WADL
document does not mention explicitly how to send such requests.