Complete CMake support for generated Endpoint proxies
Generated Endpoint proxies needs to be built together with their derived
classes. In order to do that, an additional cmake file will be generated
for each endpoint, so that it could be included and built by demand.
Endpoint base class was updated as well in order to start using new HTTP
methods and obtain JS and QML engines directly without querying them via
Session.