Fix Generator for Windows platform
Module parser was not found on Windows platform due to conflict with
the Python module with the same name. Thus, parser was renamed to
models, which is basically reflects an actual intension of modules
within that package. In addition to that, incorrect parsing of
file paths on Windows was causing errors at .cmake files, making
parts of path as escape characters.